Move existing code around to avoid mass failure
